move: 테스트 관련 utils 패키지 이동
