+++ src/test/java/com/kgu/studywithme/common/RepositoryTest.java
@@ -1,5 +1,6 @@
 package com.kgu.studywithme.common;
 
+import com.kgu.studywithme.common.annotation.MySqlTestContainers;
 import com.kgu.studywithme.global.config.P6SpyConfiguration;
 import com.kgu.studywithme.global.config.QueryDslConfiguration;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
