+++ src/test/java/com/kgu/studywithme/common/annotation/RedisTestContainers.java
@@ -1,6 +1,6 @@
-package com.kgu.studywithme.common;
+package com.kgu.studywithme.common.annotation;
 
-import com.kgu.studywithme.common.config.MySqlTestContainersConfiguration;
+import com.kgu.studywithme.common.config.RedisTestContainersConfiguration;
 import org.springframework.test.context.ContextConfiguration;
 
 import java.lang.annotation.ElementType;
@@ -10,6 +10,6 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-@ContextConfiguration(initializers = MySqlTestContainersConfiguration.Initializer.class)
-public @interface MySqlTestContainers {
+@ContextConfiguration(initializers = RedisTestContainersConfiguration.Initializer.class)
+public @interface RedisTestContainers {
 }
